Editor interface updated
